update demo and packages
