+++ source/PerformanceTests/ExamplePerformanceTests/TestBase.cs
@@ -3,11 +3,13 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.AspNetCore.TestHost;
+using PerformanceTests.DemoUtils;
+using Sailfish.AdapterUtils;
 using Test.API;
 
 namespace PerformanceTests.ExamplePerformanceTests;
 
-public class TestBase
+public class TestBase : ISailfishFixture<SailfishDependencies>
 {
     public TestBase(WebApplicationFactory<DemoApp> factory)
     {
