+++ source/PerformanceTests/PerformanceTests.csproj
@@ -11,7 +11,7 @@
         <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="6.0.5" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.1" />
         <PackageReference Include="Microsoft.TestPlatform.TestHost" Version="17.4.1" />
-        <PackageReference Include="Sailfish.TestAdapter" Version="0.0.272" />
+        <PackageReference Include="Sailfish.TestAdapter" Version="0.1.122" />
         <PackageReference Include="Serilog" Version="2.12.0" />
     </ItemGroup>
