+++ source/Sailfish/Execution/ITypeResolver.cs
@@ -5,5 +5,5 @@ namespace Sailfish.Execution;
 public interface ITypeResolver
 {
     object ResolveType(Type type);
-    T ResolveType<T>();
+    T ResolveType<T>() where T : notnull;
 }
